+++ src/main/kotlin/com/jetbrains/interactiveRebase/dataClasses/BranchInfo.kt
@@ -18,6 +18,12 @@ data class BranchInfo(
     var currentCommits: MutableList<CommitInfo> = initialCommits.toMutableList()
     var baseCommit: CommitInfo? = null
 
+    /**
+     * True if collapsed commits were expanded but were automatically collapsed again
+     * Used to enable collapse action
+     */
+    var isNestedCollapsed: Boolean = false
+
     @Synchronized
     internal fun addListener(listener: Listener) = listeners.add(listener)
 
@@ -53,17 +59,26 @@ data class BranchInfo(
 
         val collapsedCommits = this.currentCommits.subList(initialIndex, finalIndex).deepClonePolymorphic()
         val parentOfCollapsedCommit = this.currentCommits[finalIndex]
+        collapseCommitsWithList(collapsedCommits, parentOfCollapsedCommit)
+    }
 
+    /**
+     * Given a list of commits to be collapsed and a parent, collapses the commits.
+     */
+    fun collapseCommitsWithList(
+        collapsedCommits: List<CommitInfo>,
+        parentOfCollapsedCommit: CommitInfo,
+    ) {
         val collapsedCommand = CollapseCommand(parentOfCollapsedCommit, collapsedCommits.toMutableList())
 
-        parentOfCollapsedCommit.addChange(collapsedCommand)
+        parentOfCollapsedCommit.changes.add(collapsedCommand)
         parentOfCollapsedCommit.isCollapsed = true
         parentOfCollapsedCommit.isHovered = false
 
         this.currentCommits.removeAll(collapsedCommits)
         this.clearSelectedCommits()
         collapsedCommits.forEach {
-            it.addChange(collapsedCommand)
+            it.changes.add(collapsedCommand)
             it.isCollapsed = true
         }
     }